Re-re organized the midi functions tables back to happening during the processing step.
They will now be in charge of converting data to instructions ready for music_player.lua

The overall desision to switch back to this method was: If I had a table of functions to organize data during read, I'd still need a table to organize them during process anyways. So lets make the read step more generic, and add just a little bit of logic to make it all work during processing. Now, only one set of tables.