placing the sound exactly at eye level causes a terible wobble effect while the player is in motion as the game can't decide weather to place the sound in the left or right ear.
Removing `client:getFrameTime()` fixes the rapid wobble, but doesn't
solve the one-ear-or-the-other issue.

Moving the source from the eye/ear position to somewhere else (like the
center of the bounding box) solves almost all of the wobble, but it does
mean the music is slightly louder if the player looks straight down.