Boosted volume of triangle_sine.ogg
Previously, ABC files could only play max volume, and the "trick" to
getting dynamics was to play a note on top of another note. This would
get very loud, so to accomidate that, I artificialy made the triangle
audio to be quieter.

Midi is smarter and so this "doubling" isn't needed, but also we're
going to be competing with Minecraft's built in sounds, and eventualy
Figura Piano, which are louder by default.