Added a way to reduce the temporal resolution of modifiers.
If a modifier is within 20 miliseconds of the previously added modifier,
it won't be added. (Unless it the final modifier, or it is over 60
miliseconds away from the upcomming modifier. Then the modifier is
considered a "bookend" and will be added anyways.)

Anoyingly, if the pan modifier is enabled, this change doesn't help Wii
Sports much (it only shaves about 1 second of buffer time). Wii Sports
assigns `pan` to almost every channel note in the song, so every note
still has extra data about it.

To do this properly, I would have to actualy keep track of channels
during playback instead of duplicating modifiers for every note.

Conviniently though, this update removes the buffer time entirely when
the pan modifier is disabled, so this update does appear to successfuly
reduce the resolution.