Added a fake delay before playback.
The serializer is getting a little heavy, and the game's tick clock is
starting to outrun it. This means that by the time we can actualy hear
the song, the song update loop is already trying to catch up with the
client's time. This fake delay lets the game stabalize a little first.