Adjusted ping_loop logic to remove a source of "slip" when timing when to send the next packet.
If we miss the ideal window to send a packet using a TICK event, we have
to wait (worst case) 50 miliseconds for the next tick. Using the
previous solution, that 50 milisecond slip could add up over the course
of several packets, allowing the song to catch up.

This change ensures packets are sent on average
min_milis_between_packets, and allows the ping_loop to catch up before
it falls behind.