surpressed some print statements behind debug flags.
Tested in multiplayer. The receive packet system is just a little too
heavy for the default permission level (limit of ~8000 instructions). We
should try to some sort of processing loop to ease up on instructions in
a single tick. (Especialy if we're sharing instructions with other
scripts.) Works great on "high," however.