Fixed a bug where the UI might assume a song has ended before it starts if the song's header/config info was split into two packets.
Technicaly the UI was correct, since the config packet is the one that
includes the rule to start the song, the song wouldn't have started
until that packet arrives.

This is sort of a bandaid patch, since the fix is to just "wait until
we've at least had the opprotunity to hear 3 packets", so the UI now
might think songs shorter than 600ms (rare) are still playing until it
hits the 600ms mark.

Probably a worthwhile tradeoff.